requirements.txt
flask>=3.0.0
pymupdf>=1.24.0
ebooklib>=0.18
reportlab>=4.0.0
pytesseract>=0.3.13
pillow>=9.0.0
